adds a description of /api/articles/:article_id  endpoint to /api endpoint; renames files for consistency
