+++ app.js
@@ -1,19 +1,18 @@
-const { allTopics } = require("./controllers/topics.controllers");
+const { getTopics } = require("./controllers/getTopics.controllers");
 const { allEndpoints } = require("./controllers/endpoints.controllers");
-const { getArticleById } = require("./controllers/getArticle.controllers");
+const { getArticleById } = require("./controllers/getArticleById.controllers");
 const express = require("express");
 
 const app = express();
 
 app.use(express.json());
 
-app.get("/api/topics", allTopics);
+app.get("/api/topics", getTopics);
 app.get("/api", allEndpoints);
 app.get("/api/articles/:article_id", getArticleById);
 
 app.use((err, req, res, next) => {
   if (err.code === "22P02") {
-    console.log(err.code);
     res.status(400).send({ msg: "Bad Request" });
   } else {
     next(err);
