+++ controllers/getTopics.controllers.js
@@ -1,0 +1,11 @@
+const { setTopics } = require("../models/setTopics.models");
+
+exports.getTopics = (req, res, next) => {
+  setTopics()
+    .then((topics) => {
+      res.status(200).send({ topics });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
