+++ models/setTopics.models.js
@@ -1,6 +1,6 @@
 const db = require("../db/connection");
 
-exports.getTopics = () => {
+exports.setTopics = () => {
   return db.query(`SELECT * FROM topics`).then(({ rows }) => {
     return rows;
   });
